requirements.txt
playwright==1.41.0
google-generativeai==0.3.2
flask==3.0.0
pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0

